// src/modules/API/Swagger/ecotrust/gen/models/PriceItem.ts
/* eslint-disable @typescript-eslint/no-unused-vars */
/* eslint-disable @typescript-eslint/ban-ts-comment */
// @ts-nocheck
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { CurrencyInfo } from './CurrencyInfo.ts'
import type { PriceSource } from './PriceSource.ts'

export type PriceItem = {
  /**
   * @description Trading pair symbol (e.g., USD-IRT, USDT-IRT)
   * @type string
   */
  symbol: string
  /**
   * @description Source of the price data
   * @type object
   */
  price_source: PriceSource
  /**
   * @type object
   */
  name: CurrencyInfo
  /**
   * @type object
   */
  base_currency: CurrencyInfo
  /**
   * @type object
   */
  quote_currency: CurrencyInfo
  /**
   * @description URL to asset icon/logo
   * @type string, uri
   */
  png?: string | null
  /**
   * @description Sell price in quote currency (decimal string)
   * @type string
   */
  sell_price: string
  /**
   * @description Buy price in quote currency (decimal string)
   * @type string
   */
  buy_price: string
  /**
   * @description Whether the asset is tradable
   * @type boolean
   */
  is_tradable: boolean
  /**
   * @description Support price level
   * @type string
   */
  support?: string | null
  /**
   * @description Resistance price level
   * @type string
   */
  resistance?: string | null
  /**
   * @description Buy fee percentage
   * @type string | undefined
   */
  buy_fee?: string
  /**
   * @description Sell fee percentage
   * @type string | undefined
   */
  sell_fee?: string
  /**
   * @description Price change percentage
   * @type string | undefined
   */
  change?: string
  /**
   * @description Bubble value (difference from real value)
   * @type string
   */
  bubble?: string | null
  /**
   * @description Real/underlying value
   * @type string
   */
  real_value?: string | null
  /**
   * @description Bubble as percentage
   * @type string
   */
  bubble_percentage?: string | null
  /**
   * @description Last update timestamp (ISO 8601)
   * @type string, date-time
   */
  created_at: string
  /**
   * @description Whether the price is considered up-to-date
   * @type boolean
   */
  is_up_to_date: boolean
}