// src/modules/API/Swagger/ecotrust/gen/models/PricesResponse.ts
/* eslint-disable @typescript-eslint/no-unused-vars */
/* eslint-disable @typescript-eslint/ban-ts-comment */
// @ts-nocheck
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { PriceItem } from './PriceItem.ts'

export type PricesResponse = {
  /**
   * @description Array of price items
   * @type array
   */
  data: PriceItem[]
}